Project 5 Complete
Uses Graph data structure to store dependencies between rules. Rule evaluation
is optimized by finding the Strongly Connected Components of the graph and
evaluating them in topological order.

+++ Relation.h
@@ -151,7 +151,7 @@ class Relation {
             for (unsigned i = 0; i < right.scheme.size(); i++) {
                 bool unique = true;
                 for (const auto& pair : matchingColumns) {
-                    if (pair.second == i) {
+                    if (static_cast<unsigned int>(pair.second) == i) {
                         unique = false;
                     }
                 }
